Updated. Deleting files older than 1 day

+++ tx_remove_hs_err_pid_log/main.py
@@ -4,6 +4,7 @@
 __author__ = 'ipetrash'
 
 
+import datetime as DT
 import time
 
 from pathlib import Path
@@ -14,11 +15,18 @@ DIRS = [r'C:\DEV__TX', r'C:\DEV__OPTT', r'C:\DEV__RADIX']
 
 
 def run(dirs: list[Union[str, Path]]):
+    print(f'\n{DT.datetime.today()}')
+
     for dir_path in dirs:
         print(dir_path)
         for file_name in Path(dir_path).glob('*/hs_err_pid*.log'):
-            print(file_name)
-            file_name.unlink(missing_ok=True)
+            ctime_timestamp = file_name.stat().st_ctime
+            ctime = DT.datetime.fromtimestamp(ctime_timestamp)
+            print(f'{file_name} (date creation: {ctime})')
+
+            # Удаление, если с даты создания прошло больше 1 дня
+            if DT.datetime.today() > ctime + DT.timedelta(days=1):
+                file_name.unlink(missing_ok=True)
 
 
 if __name__ == '__main__':
